+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Add `toc.Transformer` to generate a table of contents to the front of any
+  document parsed by a Goldmark parser.
+- Add `toc.Extender` to extend a `goldmark.Markdown` object with the
+  transformer.
+
 ## [0.1.0] - 2021-03-23
 - Initial release.